split worker into different files
